move invert4 to util file
